requirements.txt
allure-pytest==2.9.45
allure-python-commons==2.9.45
altgraph==0.17
atomicwrites==1.4.0
attrs==21.4.0
colorama==0.4.4
distlib==0.3.4
filelock==3.6.0
future==0.18.2
iniconfig==1.1.1
MouseInfo==0.1.3
packaging==21.3
pefile==2021.5.24
platformdirs==2.5.1
pluggy==1.0.0
prettytable==2.2.0
py==1.11.0
PyAutoGUI==0.9.53
PyGetWindow==0.0.9
pyinstaller==4.3
pyinstaller-hooks-contrib==2021.1
PyMsgBox==1.0.9
pyparsing==3.0.7
pyperclip==1.8.2
pypiwin32==223
PyRect==0.1.4
PyScreeze==0.1.27
pytest==7.1.0
PyTweening==1.0.3
pywin32==301
pywin32-ctypes==0.2.0
six==1.16.0
tomli==2.0.1
virtualenv==20.14.0
wcwidth==0.2.5
